* 'md4_cleanup' of git://github.com/krajaratnam/ccache:
  use size_t when we refer to memory sizes
  Remove unused 'void mdfour(unsigned char *out, const unsigned char *in, int n)'
  Remove incorrect assumption
  Remove disabled md4 test code

+++ unify.c
@@ -102,7 +102,7 @@ static void build_table(void)
 static void pushchar(struct mdfour *hash, unsigned char c)
 {
 	static unsigned char buf[64];
-	static int len;
+	static size_t len;
 
 	if (c == 0) {
 		if (len > 0) {
